refactor(shared): move webview panel chrome from modeler-types to shared
The #1371 split placed the resizer/focus/shortcut chrome in the publishable
types package on the no-protocol-dependency criterion alone. The files are
page-level webview chrome bound to hard-coded DOM ids and marked @internal,
yet exported from the publishable index. Move them to the private shared lib
while the package is unpublished and the move is still free; only the pure
isTextEditingSurface predicate stays public. See ADR 0019.
