+++ libs/shared/src/index.ts
@@ -4,4 +4,6 @@ export * from "./lib/messages";
 export * from "./lib/documentFlush";
 export * from "./lib/modeler";
 export * from "./lib/processVariables";
+export * from "./lib/propertiesPanelResizer";
+export * from "./lib/propertiesPanelFocus";
 export * from "./lib/variableManifest";
