+++ libs/shared/src/lib/propertiesPanelFocus.spec.ts
@@ -5,7 +5,6 @@ import type { PropertiesPanelHandle } from "./propertiesPanelResizer";
 import {
     focusPropertiesPanel,
     installPanelShortcuts,
-    isTextEditingSurface,
     togglePropertiesPanel,
 } from "./propertiesPanelFocus";
 import type { PanelFocusOptions } from "./propertiesPanelFocus";
@@ -47,40 +46,6 @@ function createPanel(...fieldTags: string[]): { panel: HTMLDivElement; opts: Pan
 function cleanupPanels(): void {
     document.querySelectorAll("#js-properties-panel").forEach((el) => el.remove());
 }
-
-// ────────────────────────────────────────────────────────────────────
-// isTextEditingSurface
-// ────────────────────────────────────────────────────────────────────
-
-describe("isTextEditingSurface", () => {
-    it("returns true for <input>", () => {
-        expect(isTextEditingSurface(document.createElement("input"))).toBe(true);
-    });
-
-    it("returns true for <textarea>", () => {
-        expect(isTextEditingSurface(document.createElement("textarea"))).toBe(true);
-    });
-
-    it("returns true for contenteditable element", () => {
-        const div = document.createElement("div");
-        div.contentEditable = "true";
-        expect(isTextEditingSurface(div)).toBe(true);
-    });
-
-    it("returns false for a plain <div>", () => {
-        expect(isTextEditingSurface(document.createElement("div"))).toBe(false);
-    });
-
-    it("returns false for null", () => {
-        expect(isTextEditingSurface(null)).toBe(false);
-    });
-
-    it("returns false for contenteditable=inherit", () => {
-        const div = document.createElement("div");
-        div.contentEditable = "inherit";
-        expect(isTextEditingSurface(div)).toBe(false);
-    });
-});
 
 // ────────────────────────────────────────────────────────────────────
 // focusPropertiesPanel
