+++ libs/shared/src/lib/propertiesPanelFocus.ts
@@ -1,20 +1,10 @@
 /**
- * @internal Page-level webview chrome coupled to the `js-properties-panel` DOM
- * id and the panel shortcuts. Not part of the modeler public API.
- * `isTextEditingSurface` is the one broadly-reused helper here.
+ * Page-level webview chrome coupled to the `js-properties-panel` DOM id and
+ * the panel shortcuts. Not part of the modeler public API.
  */
-import type { PropertiesPanelHandle } from "./propertiesPanelResizer";
+import { isTextEditingSurface } from "@miragon/bpmn-modeler-types";
 
-/**
- * True when the element is a text-editing surface where single-character
- * keystrokes must type rather than trigger shortcuts.
- */
-export function isTextEditingSurface(el: Element | null): boolean {
-    if (el instanceof HTMLInputElement) return true;
-    if (el instanceof HTMLTextAreaElement) return true;
-    if (el instanceof HTMLElement && el.contentEditable === "true") return true;
-    return false;
-}
+import type { PropertiesPanelHandle } from "./propertiesPanelResizer";
 
 export interface PanelFocusOptions {
     /** Override the panel root lookup (default: `getElementById("js-properties-panel")`). */
