+++ libs/shared/src/lib/propertiesPanelResizer.ts
