+++ libs/shared/vitest.config.ts
@@ -1,10 +1,14 @@
 import { defineConfig } from "vitest/config";
+import { resolve } from "node:path";
 
 export default defineConfig({
     test: {
         name: "shared",
         environment: "node",
         include: ["src/**/*.{spec,test}.ts"],
+        alias: {
+            "@miragon/bpmn-modeler-types": resolve(__dirname, "../modeler-types/src/index.ts"),
+        },
         coverage: {
             provider: "v8",
             reportsDirectory: "../../coverage/libs/shared",
